Fixes formatting error
This fixes a runtime error due to incorrect value/type being supplied. Realistically it should be determined whether 7hm's version of this file or mine should be kept, and discard the alternate file. It seems as though my file is the one in the "Random Collection of Zones" archive folder.